Add masonry to layout for card rearranging
Use masonry and jQuery to get a dynamic layout for the cards on the
items index and category show pages.
There are 3 columns in the large layout, 2 in the medium, and 1 in
the small. Columns are always centered and are not arranged until
images have been loaded. Card titles are also 80% opaque.