Delete item from cart
User can delete an item from the cart by clicking on the 'Remove'
button next to the item. A flash message will be displayed that also
allows the user to go to the show page for the item they just deleted
in case they want to add the item again. The flash message is colored
green since deleting items is a success message.