Move cart item to partial view
The cart index now uses a partial to render each item in the cart.
Additionally, within this view, the quantity and price are now
on the same line. There is an additional class added to quantity so
that it is 20pixels away from the price.

Also removes the valign because this allows the view to look better
when the screen is mobile-size by moving the text below the images.
We may want to consider dynamically adding valign classes with jquery
for larger screen sizes.