Turn item card image into a link to item show page
Changes the 'div' tag to be an 'a' tag for the item image in the
item_card partial. Adds a link to go to the specific item's show page
so that when you click on the card's image you can go directly to its
page to see all of the other details.