User can edit item quantity in cart
Item quantity in the cart is an editable field linked to a form
and an update button so the quantity can be changed. The total number
of items in the cart is also displayed next to the shopping cart icon.
Note this branch is called 'model test' but that will actually be
worked on on a different branch.