Creates user login tests
Tests for user logging in, unauthorized user attempting to log in,
and user logging in with incorrect username or password are now
written and passing. Additionally, the SessionsController is now
correctly inheriting from ApplicationController. It was inheriting
from some other class before which is why the application.html.erb
file wasn't being used as the layout. This is now fixed, so the
navbar and other styling is visible on the page.