Order index uses order_card partial
Updated the way that the order index page and order show page pass
variables to the order_card partial so that they can both use the partial.
This seems to have fixed the subtotal problem that Brenna mentioned.
Still need to work on the responsiveness of the order cards for small
screen sizes.