Add check for empty cart before placing order
Adds a method to the cart class to check if the contents of the
cart is empty. If it is, the user will be redirected to the cart
and a flash message will be displayed saying that an order cannot
be placed with no items. All the tests are still passing.