Create Artists scope on User and artists index
Allows a user to browse art listed in an index by artist. Creates
a 'artists' scope on the User class to be able to grab all users
who are artists. Also adds a full_name method to the User class which
is a string of their first and last name with a space inbetween them.
Adds a test for browsing by artist.