Add model tests for Item, add required form fields
Adds models tests for the Item class to allow for a default photo
to be used if one is not provided. Also adds required: true to some
of the form fields for admin new item creation. Modifies the navbar
to show if the current user is an admin. Cleans up the admin new
item creation test to use factory girl. Updates the seeds.rb file
to create both artists and users.