Guest can create artist account
The create account screen now has a dropdown to select what kind of
user (default or artist) the user would like to sign up as. Admin is
not an option, except when an admin is editing their own information
when logged in, in which case it defaults to keeping them an admin.
An artist will see links for viewing their items and adding a new item
from their dashboard.