Add delete to index, add negative path test
Adds the delete button to the index (_item_card) partial since the
edit button is there also. Adds a negative path test for an artist not
being able to delete another artist's item. Adds the before action
to check to ensure the item is the current user's item to the
destroy action in the ItemsController.